Remove contact-types.ts
