+++ shared-libs/cht-datasource/test/contact.spec.ts
@@ -5,7 +5,6 @@ import * as Local from '../src/local';
 import * as Remote from '../src/remote';
 import * as Qualifier from '../src/qualifier';
 import * as Contact from '../src/contact';
-import * as ContactType from '../src/contact-types';
 import { expect } from 'chai';
 import * as Core from '../src/libs/core';
 
@@ -28,31 +27,6 @@ describe('contact', () => {
   afterEach(() => sinon.restore());
 
   describe('v1', () => {
-    describe('isNormalizedParent', () => {
-      let isDataObject: SinonStub;
-
-      beforeEach(() => isDataObject = sinon.stub(Core, 'isDataObject'));
-      afterEach(() => sinon.restore());
-
-      ([
-        [{ _id: 'my-id' }, true, true],
-        [{ _id: 'my-id' }, false, false],
-        [{ hello: 'my-id' }, true, false],
-        [{ _id: 1 }, true, false],
-        [{ _id: 'my-id', parent: 'hello' }, true, false],
-        [{ _id: 'my-id', parent: null }, true, true],
-        [{ _id: 'my-id', parent: { hello: 'world' } }, true, false],
-        [{ _id: 'my-id', parent: { _id: 'parent-id' } }, true, true],
-        [{ _id: 'my-id', parent: { _id: 'parent-id', parent: { hello: 'world' } } }, true, false],
-        [{ _id: 'my-id', parent: { _id: 'parent-id', parent: { _id: 'grandparent-id' } } }, true, true],
-      ] as [unknown, boolean, boolean][]).forEach(([value, dataObj, expected]) => {
-        it(`evaluates ${JSON.stringify(value)}`, () => {
-          isDataObject.returns(dataObj);
-          expect(ContactType.v1.isNormalizedParent(value)).to.equal(expected);
-        });
-      });
-    });
-
     describe('get', () => {
       const contact = { _id: 'my-contact' } as Contact.v1.Contact;
       const qualifier = { uuid: contact._id } as const;
